Add ALT key press handling
ECMA-35 and a host of terminals send ESC followed by a keypress as
a 7-bit encoding for representing that ALT (or META) was held down
for that keypress.

We need to keep some more state in the input parser to track this.
There's a little bit of nuance; we don't want to block processing
of a lone escape indefinitely.

The keymap code has been changed to return a clone of the value
in lookup to appease the borrow checker: it can't see that we don't
modify the keymap in the context of the dispatch_callback function :-/