fix rendering of reverse video
This got broken when unifying the termwiz color type; we used to
have enum variants for background and foreground colors, but now
we just have Default and we need to resolve using helper methods
in the palette.

That means we need to resolve the colors before we swap them for
reverse video.

Bonus: we can lift this out of the loop for the cell cluster.