whoops, fully undo the half-hearted attempt to address the todo in here
The saturating_sub was going to be an attempt at the look-back alluded
to in the TODO comment, but I decided not to implement that in this
iteration.

The effect of this was that double clicking the `m` in a sequence like
` master` would only select the `m` instead of the `master` string.