Remove perssimstic do_paint call
This do_paint call was added in 8995974124964ce45935d996da415a9a03f5e977
but I don't think it was intentional; it looks like an accidental commit
of something during debugging.

The intent is that we tick over and paint every tick (~50ms) and this
behavior here was causing us to paint after every input event, which is
undesirable when processing the selection.

refs: https://github.com/wez/wezterm/issues/33