Improve FontKit rendering, but it is still bad
I don't think font-kit has enough data available for to make things
work the way we're used to.  In addition, the implementation of
the loader for core text has a lot of TODO's.

The state of the FontKit variant is that it basically renders glyphs
and they are aligned to the same baseline, but both the vert and
horz metrics are off; for example, `-` doesn't render.  The glyphs
are also under-sized.