fonts: fixed up shaping using allsorts
The issue mentioned from the prior commit was due to not breaking
out of the inner fallback loop, so we'd stack up multiple glyphs
with the same cluster value, causing the earlier versions to be
obscured by the later versions.

Adjust glyph width calculation: the metrics we were getting from
harfbuzz were synchronized with the font render size.  When we're
configured to use allsorts, the shaper metrics are not connected
to the render metrics and we can end up with the raw emoji glyph
width being much larger than the font advance metric value and
then render a giant heart emoji.   The revised calculations trigger
scaling if the glyph width is too wide.  I've gone back and forth
on that particular line a couple of times in the past: hopefully
this time we've got the right calculation?

I "fixed" the font metrics we compute and return from allsorts
by rounding them to the nearest integer value.  That makes the
spacing look better at my normal font size (8.3 -> 8) and makes
things look tighter.  It feels a bit magical IMO.