fonts: readjust the glyph scaling code
As mentioned in f204ad9a827e43434e8f3e06e1a37ba035392da2, this has
gone back and forth a few times.

This version avoids some artifacts by avoiding scaling in most cases.
The test scenario for this is to tab complete a directory name in
zsh; that causes a bold `/` glyph to be rendered which selects a
typeface with different metrics and would render a horizontal line
at either the top of bottom of the glyph.

Similarly, a `/` in italics (eg: comments in vim) would select a
third different font and have different artifact properties.

Now both of those artifacts are eliminated.