fix toml parse error reporting
The move to anyhow changed the nature of the error objects
that get passed through to the notification system; since
we use the `context` method and used the normal display
presentation (via `to_string`) we were only reporting the
context and not the nature of the error.

Switching to alternate output makes the error messages more
useful by including the underlying problem and line number.

Refs: https://github.com/wez/wezterm/issues/78