catch panic -> toast notification
This is useful when launched from a gui launcher and the stderr is
not conveniently visible.

If a panic occurs, generate a toast notification with the panic
message so that there is a breadcrumb to follow if the application
does terminate in this way.

Every so often I encounter this situation on my windows machine during
a graphics card driver update, so this should help to run that down.