cargo update to pick up filedescriptor 0.7.1
I got to the bottom of the hang on startup when trying to connect
to the multiplexer on windows and it was because of the change
in 367a44fb96d9ebd4a38621cc6afdde04a320350c to try to drain the pollable
channel.

This mapped to ReadFile on windows which always tries to fill the buffer
even if it is a socket under the covers.

https://github.com/wez/wzsh/commit/348421b0109f3be498dc23fc3937b5fba7c97132
teaches filedescriptor to use the recv/send if the filedescriptor is
really a socket and that resolves the hang.