lua: keys: add some action plumbing
This commit tackles a couple of things in related areas:

Makes it possible to reference the KeyAssignment type directly from the
KeyAction action type.  Previously, due to limitations in how TOML is
serialized, the enum values could be used here; while the compiler would allow
it, it is impossible to construct a TOML value for this sort of enum.  Now that
we have lua we have more flexibility in the table type and this is unblocked.
In practical terms this means that we can do away with the janky `arg` field
and just use the `action` field directly, which also means that we can do away
with a whole extra enum that duplicates just the enum variants and makes
maintenance easier!

Adds `wezterm.action` to lua which helps to construct these enum variants.  The
helper is used like this in the `.wezterm.lua` file:

```lua
local wezterm = require 'wezterm';
return {
   keys = {
     {key="{", mods="SHIFT|CTRL", action=wezterm.action{ActivateTabRelative=-1}},
     {key="}", mods="SHIFT|CTRL", action=wezterm.action{ActivateTabRelative=1}},
   }
}
```

the above creates values like `KeyAssignment::ActivateTabRelative(-1)`
in a not totally awful syntax.

On top of this, which was the main reason for doing this work, this
commit adds two new variants to KeyAssignment that are accessible
on from lua (because of the TOML limitations mentioned above);
`SpawnCommandInNewWindow` and `SpawnCommandInNewTab` that take a
`SpawnCommand` struct as their parameter.  These actions are
conceptually similar to the existing `SpawnTab` key assignment
but allow specifying the program to be run as well as its cwd
and environment.

For example, this key assignment spawns `top` in a new window:

```lua
  {key="y", mods="CMD", action=wezterm.action{SpawnCommandInNewWindow={
    args={"top"}
  }}},
```

The intent is to use this as a building block for saved/canned
commands.

refs: https://github.com/wez/wezterm/issues/159