mux: set a limit on the line cache
I made this unbounded originally because I wanted to do something
smarter around config reloading.  I'm still too lazy to make it
do that, but do recognize that we should have some limit on the
line cache size, so we just pick the value of the scrollback
size at the time we started.