mux: add basic "predictive" local echo
This is a mosh-like feature, but not as sophisticated.

The goal is to reduce perceived latency by making a reasonable
guess about how the line where the cursor is located will change
in response to a key press.

At the moment the guess has some very basic heuristics:

* The key is an unmodified (no CTRL, ALT, SUPER) `char` value
* The line containing the cursor doesn't appear to be a
  password prompt (using a very simple English specific test)

In those conditions, we'll update the local line cache and cursor
position with how we think the text will appear, but we set the
underline attribute (which is relatively rarely used) to make it clearer
that this is synthetic.

For low latency environments (eg: local network) the underline
rarely appears unless you are able to type very quickly.

I've yet to test this on a higher latency link, but think
this is good enough to land and build out so that I can test
that scenario with my work systems.

refs: https://github.com/wez/wezterm/issues/127