mux: add input event serial number to key events and render responses
This should allow some basic reasoning about how in sync we are with
the remote system.

The immediate application of this is to try to avoid wiggling the
text cursor when the predictive echo updates the position locally.

I need to push this in order to test it on my higher latency setup,
so I don't know if this is totally effective yet.